move api modules into /api/ directory, don't import public api in private modules
