+++ src/sciform/__init__.py
@@ -1,8 +1,8 @@
 """``sciform`` is used to convert python numbers into scientific formatted strings."""
 
-from sciform.formatter import Formatter
+from sciform.api.formatter import Formatter
 from sciform.formatting import FormattedNumber
-from sciform.global_configuration import (
+from sciform.api.global_configuration import (
     GlobalOptionsContext,
     get_default_global_options,
     get_global_options,
@@ -12,7 +12,7 @@ from sciform.global_configuration import (
 from sciform.modes import AutoDigits, AutoExpVal
 from sciform.options.input_options import InputOptions
 from sciform.options.populated_options import PopulatedOptions
-from sciform.scinum import SciNum
+from sciform.api.scinum import SciNum
 
 __all__ = [
     "Formatter",
