+++ src/sciform/api/global_configuration.py
