+++ tests/test_docstrings.py
@@ -1,6 +1,7 @@
 import doctest
 
-from sciform import formatter, output_conversion, scinum
+from sciform import output_conversion
+from sciform.api import formatter, scinum
 from sciform.options import input_options, populated_options
